fixed quotes removal, better call saul, shell variables are backgit add -u! and juicier than ever

+++ src/vars/mtxdb/mtx_move.c
@@ -6,7 +6,7 @@
 /*   By: topiana- <topiana-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/26 21:28:55 by topiana-          #+#    #+#             */
-/*   Updated: 2025/03/26 21:38:20 by topiana-         ###   ########.fr       */
+/*   Updated: 2025/04/02 10:47:08 by topiana-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@ void	mtx_move(int src, int dest, char **mtx)
 {
 	if (mtx == NULL)
 		return ;
-	if (src < 1 || src >= *(int *)mtx
-		|| dest < 1 || dest >= *(int *)mtx )
+	if (src < 1 || (unsigned int)src >= *(unsigned int *)mtx
+		|| dest < 1 || (unsigned int)dest >= *(unsigned int *)mtx )
 		return ;
 	mtx[dest] = mtx[src];
 	mtx[src] = NULL;
@@ -28,19 +28,19 @@ void	mtx_move(int src, int dest, char **mtx)
 
 void	mtx_pushup(char **mtx)
 {
-	int	i[2];
+	unsigned int	i[2];
 	
 	if (mtx == NULL)
 		return ;
 	ft_memset(i, 0, 2 * sizeof(int));
-	while (i[0] < *(int *)mtx)
+	while (i[0] < *(unsigned int *)mtx)
 	{
 		if (mtx[i[0]] == NULL)
 		{
 			i[1] = i[0];
-			while (i[1] < *(int *)mtx && mtx[i[1]] == NULL)
+			while (i[1] < *(unsigned int *)mtx && mtx[i[1]] == NULL)
 				i[1]++;
-			if (i[1] < *(int *)mtx)
+			if (i[1] < *(unsigned int *)mtx)
 				mtx_move(i[1], i[0], mtx);
 		}
 		i[0]++;
